+++ src/prim.h
@@ -1,13 +1,13 @@
 #ifndef PRIM_H_
 #define PRIM_H_
 
-#include <stdint.h>
+
 #include "graph.h"
 #include "min_heap.h"
 
 
 
-Graph* prim(Graph *graph, uint16_t size);
+Graph* prim(Graph *graph, int size);
 
 
 #endif /* PRIM_H_ */
